Allow modification of 'incomplete-dir' location
Sometimes it is necessary to change not only location of 'destination'
but also location of 'incomplete-dir' where temporary downloads are
stored as configured by Transmissions 'resume' files.

Added 2 options:
  -d or --destination to change torrent's destination location
  -i or --incomplete-dir to change torrents incomplete location
Now it is mandantory to specify one of them when running the script.
The logic was tested and checked with shellcheck.

README and AUTHORS are updated accordingly.